add comment entity, dto, controller and some other topic fix

+++ backend/src/main/java/com/proj/forum/dto/StatisticDto.java
@@ -11,7 +11,7 @@ public record StatisticDto(
         @Id UUID id,
         @NotNull UUID userId,
         @NotNull UUID objectId,
-        int rate,
+        Short rate,
         Boolean liked,
         @NotNull Boolean saved
         ) { }
